+++ finagle-netty4/src/main/scala/com/twitter/finagle/netty4/threading/EventLoopGroupTrackingRunnable.scala
@@ -1,15 +1,23 @@
 package com.twitter.finagle.netty4.threading
 
 import com.twitter.finagle.stats.Stat
+import com.twitter.finagle.stats.StatsReceiver
 import com.twitter.logging.Logger
-import com.twitter.util.{Duration, Time}
+import com.twitter.util.Duration
+import com.twitter.util.Time
+import io.netty.channel.SingleThreadEventLoop
 import io.netty.util.concurrent.EventExecutor
-import java.util.concurrent.{Callable, ScheduledFuture, ScheduledThreadPoolExecutor, TimeUnit}
+import java.lang.management.ManagementFactory
+import java.util.concurrent.Callable
+import java.util.concurrent.ScheduledFuture
+import java.util.concurrent.ScheduledThreadPoolExecutor
+import java.util.concurrent.TimeUnit
 
-private[threading] class EventLoopGroupExecutionDelayTrackingRunnable(
-  eventExecutor: EventExecutor,
-  injectionPeriod: Duration,
+private[threading] class EventLoopGroupTrackingRunnable(
+  executor: EventExecutor,
+  taskTrackingPeriod: Duration,
   delayStat: Stat,
+  statsReceiver: StatsReceiver,
   threadDumpThreshold: Duration,
   dumpWatchThreadPool: Option[ScheduledThreadPoolExecutor],
   dumpLogger: Logger)
@@ -22,10 +30,10 @@ private[threading] class EventLoopGroupExecutionDelayTrackingRunnable(
   // the one thread in the executor. This is currently how netty is implemented
   // but this class will stop working if netty changes their implementation
   private[this] val executorThread: Thread = {
-    if (eventExecutor.inEventLoop()) {
+    if (executor.inEventLoop()) {
       Thread.currentThread()
     } else {
-      eventExecutor
+      executor
         .submit(new Callable[Thread] {
           override def call(): Thread = {
             Thread.currentThread()
@@ -34,15 +42,26 @@ private[threading] class EventLoopGroupExecutionDelayTrackingRunnable(
     }
   }
 
+  private[this] val threadId = executorThread.getId
   private[this] val threadName: String = executorThread.getName
+
   private[this] var scheduledExecutionTime: Time = Time.now
   private[this] var watchTask: Option[ScheduledFuture[_]] = None
 
+  private[this] val threadMXBean = ManagementFactory.getThreadMXBean
+
+  private[this] val scopedStatsReceiver = statsReceiver.scope(threadName)
+  private[this] val activeSocketsStat = scopedStatsReceiver.stat("active_sockets")
+  private[this] val cpuTimeCounter = scopedStatsReceiver.counter("cpu_time_ms")
+
+  // Accessed only from within the same netty thread
+  private[this] var prevCPUTimeMs = 0L
+
   setWatchTask()
-  eventExecutor.scheduleWithFixedDelay(
+  executor.scheduleWithFixedDelay(
     this,
     0,
-    injectionPeriod.inMillis,
+    taskTrackingPeriod.inMillis,
     java.util.concurrent.TimeUnit.MILLISECONDS
   )
 
@@ -57,12 +76,26 @@ private[threading] class EventLoopGroupExecutionDelayTrackingRunnable(
       dumpLogger.warning(
         s"THREAD: $threadName EXECUTION DELAY is greater than ${threadDumpThreshold.inMillis}ms, was ${executionDelay.inMillis}ms"
       )
-
     }
 
     delayStat.add(executionDelay.inMillis)
-    scheduledExecutionTime = Time.now.plus(injectionPeriod)
+    scheduledExecutionTime = Time.now.plus(taskTrackingPeriod)
     setWatchTask()
+
+    var numActiveSockets = 0
+    // This will be nio event loop or epoll event loop.
+    executor.asInstanceOf[SingleThreadEventLoop].registeredChannelsIterator().forEachRemaining {
+      channel =>
+        if (channel.isActive) {
+          numActiveSockets += 1
+        }
+    }
+    activeSocketsStat.add(numActiveSockets)
+
+    // `getThreadCPUTime` returns the time in nanoseconds.
+    val currentCPUTimeMs = threadMXBean.getThreadCpuTime(threadId) / 1000000
+    cpuTimeCounter.incr(currentCPUTimeMs - prevCPUTimeMs)
+    prevCPUTimeMs = currentCPUTimeMs
   }
 
   private[this] def setWatchTask(): Unit = {
@@ -71,7 +104,7 @@ private[threading] class EventLoopGroupExecutionDelayTrackingRunnable(
         dumpWatchThreadPool.get.schedule(
           new Runnable {
             override def run(): Unit = {
-              var builder = new StringBuilder()
+              val builder = new StringBuilder()
               builder
                 .append(
                   s"THREAD: $threadName EXECUTION DELAY exceeded configured dump threshold. Thread stack trace:\n"
@@ -80,7 +113,7 @@ private[threading] class EventLoopGroupExecutionDelayTrackingRunnable(
               dumpLogger.warning(builder.toString())
             }
           },
-          (injectionPeriod + threadDumpThreshold).inMillis,
+          (taskTrackingPeriod + threadDumpThreshold).inMillis,
           TimeUnit.MILLISECONDS
         )
       )